new pause button location for 1600x900
okay, I got a new idea to make this app work with all computers but
user will need to input the pixel locations and the color of the pixels
or I can try to make this automated as well, somewhat.

-will need to have a save button that creates a textfile maybe in the
documents folder on a windows computer and store values in the textfile

-will need to add textboxes or something on the UI for user to input
these values to be stored into textfile

-when user starts up the app, will first check the textfile to see if
there are pixels and colors stored in there and then parse information
from the textfile and assign them to global variables for the night run.

this is how I will retain memory of values without having the
application open 24/7